% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myconstr.R
\name{myconstr}
\alias{myconstr}
\title{Returns the results from a t-test}
\usage{
myconstr(x, y, alpha = 0.05, paired = paired)
}
\arguments{
\item{x}{first inserted array}

\item{y}{second inserted array}

\item{alpha}{confidence level}

\item{paired}{determines if t-test is paired}
}
\description{
Takes in four parameters, two arrays of numbers, the confidence and a boolean
Uses both arrays to perform a t-test, saving the dataframe from the two arrays, level of confidence, confidence interval, and p-value
}
\examples{
# myconstr(x = array1, y = array2, alpha = 0.05, paired = TRUE)
}
