% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Rttest.R
\name{plot_Rttest}
\alias{plot_Rttest}
\title{Returns side-by-side boxplots for the two array inputs}
\usage{
plot_Rttest(obj)
}
\arguments{
\item{obj}{object containing the arrays which will be plotted}
}
\value{

}

\description{
Takes in one object, the same one returned from the constructor function
Creates a new dataframe of the arrays in the object
Graphs side-by-side boxplots for both of the arrays
}
\examples{
#plot_Rttest <- function(obj = object)
}
