% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_Rttest.R
\name{print_Rttest}
\alias{print_Rttest}
\title{Prints results from previous t-test}
\usage{
print_Rttest(obj)
}
\arguments{
\item{obj}{object containing results from previous t-test}
}
\description{
Takes in an object containing a dataframe with two arrays, the confidence level, confidence intervals, and p value
}
\examples{
# print_Rttest(obj)
}
