% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myconstr.R
\name{myconstr}
\alias{myconstr}
\title{Returns the results from a t-test}
\usage{
myconstr(x, y, alpha = 0.05)
}
\arguments{
\item{x}{first inserted array}

\item{y}{second inserted array}

\item{alpha}{confidence level}
}
\description{
Takes in three parameters, two arrays of numbers, and the confidence interval
Uses both arrays to perform a t-test, saving the dataframe from the two arrays, level of confidence, confidence interval, and p-value
}
\examples{
# myconstr(x = array1, y = array2, alpha = 0.05)
}
